# productCenter/api/shop/v1/shop_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.35.0
// source: shop/v1/shop.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationShopCreateShop = "/shop.v1.Shop/CreateShop"
const OperationShopDeleteShop = "/shop.v1.Shop/DeleteShop"
const OperationShopGetShop = "/shop.v1.Shop/GetShop"
const OperationShopListShops = "/shop.v1.Shop/ListShops"
const OperationShopUpdateShop = "/shop.v1.Shop/UpdateShop"

type ShopHTTPServer interface {
	// CreateShop 创建店铺
	CreateShop(context.Context, *CreateShopRequest) (*CreateShopResponse, error)
	// DeleteShop 删除店铺
	DeleteShop(context.Context, *DeleteShopRequest) (*DeleteShopResponse, error)
	// GetShop 查看店铺详情
	GetShop(context.Context, *GetShopRequest) (*GetShopResponse, error)
	// ListShops 分页查询店铺列表
	ListShops(context.Context, *ListShopsRequest) (*ListShopsResponse, error)
	// UpdateShop 更新店铺
	UpdateShop(context.Context, *UpdateShopRequest) (*UpdateShopResponse, error)
}

func RegisterShopHTTPServer(s *http.Server, srv ShopHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/shops", _Shop_CreateShop0_HTTP_Handler(srv))
	r.PUT("/api/v1/shops/{id}", _Shop_UpdateShop0_HTTP_Handler(srv))
	r.GET("/api/v1/shops/{id}", _Shop_GetShop0_HTTP_Handler(srv))
	r.GET("/api/v1/shops", _Shop_ListShops0_HTTP_Handler(srv))
	r.DELETE("/api/v1/shops/{id}", _Shop_DeleteShop0_HTTP_Handler(srv))
}

func _Shop_CreateShop0_HTTP_Handler(srv ShopHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateShopRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopCreateShop)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateShop(ctx, req.(*CreateShopRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateShopResponse)
		return ctx.Result(200, reply)
	}
}

func _Shop_UpdateShop0_HTTP_Handler(srv ShopHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateShopRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopUpdateShop)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateShop(ctx, req.(*UpdateShopRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateShopResponse)
		return ctx.Result(200, reply)
	}
}

func _Shop_GetShop0_HTTP_Handler(srv ShopHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetShopRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopGetShop)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetShop(ctx, req.(*GetShopRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetShopResponse)
		return ctx.Result(200, reply)
	}
}

func _Shop_ListShops0_HTTP_Handler(srv ShopHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListShopsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopListShops)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListShops(ctx, req.(*ListShopsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListShopsResponse)
		return ctx.Result(200, reply)
	}
}

func _Shop_DeleteShop0_HTTP_Handler(srv ShopHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteShopRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopDeleteShop)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteShop(ctx, req.(*DeleteShopRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteShopResponse)
		return ctx.Result(200, reply)
	}
}

type ShopHTTPClient interface {
	// CreateShop 创建店铺
	CreateShop(ctx context.Context, req *CreateShopRequest, opts ...http.CallOption) (rsp *CreateShopResponse, err error)
	// DeleteShop 删除店铺
	DeleteShop(ctx context.Context, req *DeleteShopRequest, opts ...http.CallOption) (rsp *DeleteShopResponse, err error)
	// GetShop 查看店铺详情
	GetShop(ctx context.Context, req *GetShopRequest, opts ...http.CallOption) (rsp *GetShopResponse, err error)
	// ListShops 分页查询店铺列表
	ListShops(ctx context.Context, req *ListShopsRequest, opts ...http.CallOption) (rsp *ListShopsResponse, err error)
	// UpdateShop 更新店铺
	UpdateShop(ctx context.Context, req *UpdateShopRequest, opts ...http.CallOption) (rsp *UpdateShopResponse, err error)
}

type ShopHTTPClientImpl struct {
	cc *http.Client
}

func NewShopHTTPClient(client *http.Client) ShopHTTPClient {
	return &ShopHTTPClientImpl{client}
}

// CreateShop 创建店铺
func (c *ShopHTTPClientImpl) CreateShop(ctx context.Context, in *CreateShopRequest, opts ...http.CallOption) (*CreateShopResponse, error) {
	var out CreateShopResponse
	pattern := "/api/v1/shops"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationShopCreateShop))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteShop 删除店铺
func (c *ShopHTTPClientImpl) DeleteShop(ctx context.Context, in *DeleteShopRequest, opts ...http.CallOption) (*DeleteShopResponse, error) {
	var out DeleteShopResponse
	pattern := "/api/v1/shops/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationShopDeleteShop))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetShop 查看店铺详情
func (c *ShopHTTPClientImpl) GetShop(ctx context.Context, in *GetShopRequest, opts ...http.CallOption) (*GetShopResponse, error) {
	var out GetShopResponse
	pattern := "/api/v1/shops/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationShopGetShop))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListShops 分页查询店铺列表
func (c *ShopHTTPClientImpl) ListShops(ctx context.Context, in *ListShopsRequest, opts ...http.CallOption) (*ListShopsResponse, error) {
	var out ListShopsResponse
	pattern := "/api/v1/shops"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationShopListShops))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateShop 更新店铺
func (c *ShopHTTPClientImpl) UpdateShop(ctx context.Context, in *UpdateShopRequest, opts ...http.CallOption) (*UpdateShopResponse, error) {
	var out UpdateShopResponse
	pattern := "/api/v1/shops/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationShopUpdateShop))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}