# productCenter/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/bff/shops/{id}:
        get:
            tags:
                - BFF
            description: 店铺主页 — 聚合店铺 + 商品列表
            operationId: BFF_GetShopHome
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/bff.v1.GetShopHomeResponse'
...
    /api/v1/shops:
        get:
            tags:
                - Shop
            description: 分页查询店铺列表
            operationId: Shop_ListShops
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/shop.v1.ListShopsResponse'
        post:
            tags:
                - Shop
            description: 创建店铺
            operationId: Shop_CreateShop
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/shop.v1.CreateShopRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/shop.v1.CreateShopResponse'
    /api/v1/shops/{id}:
        get:
            tags:
                - Shop
            description: 查看店铺详情
            operationId: Shop_GetShop
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/shop.v1.GetShopResponse'
        put:
            tags:
                - Shop
            description: 更新店铺
            operationId: Shop_UpdateShop
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/shop.v1.UpdateShopRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/shop.v1.UpdateShopResponse'
        delete:
            tags:
                - Shop
            description: 删除店铺
            operationId: Shop_DeleteShop
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/shop.v1.DeleteShopResponse'
...
components:
    schemas:
...
        bff.v1.GetShopHomeResponse:
            type: object
            properties:
                shop:
                    $ref: '#/components/schemas/bff.v1.ShopHome'
...
        bff.v1.ProductListItem:
            type: object
            properties:
                id:
                    type: string
                shopId:
                    type: string
                shopName:
                    type: string
                name:
                    type: string
                type:
                    type: string
                mainImageUrl:
                    type: string
                price:
                    type: string
                compareAtPrice:
                    type: string
                status:
                    type: integer
                    format: int32
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.Tag'
                createdAt:
                    type: string
            description: 商品列表项（精简版）
        bff.v1.ShopHome:
            type: object
            properties:
                id:
                    type: string
                shopName:
                    type: string
                description:
                    type: string
                products:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.ProductListItem'
            description: 店铺主页数据
...
        bff.v1.Tag:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
...
        shop.v1.CreateShopRequest:
            type: object
            properties:
                shopName:
                    type: string
                description:
                    type: string
            description: 创建店铺
        shop.v1.CreateShopResponse:
            type: object
            properties:
                shop:
                    $ref: '#/components/schemas/shop.v1.ShopInfo'
        shop.v1.DeleteShopResponse:
            type: object
            properties:
                success:
                    type: boolean
        shop.v1.GetShopResponse:
            type: object
            properties:
                shop:
                    $ref: '#/components/schemas/shop.v1.ShopInfo'
        shop.v1.ListShopsResponse:
            type: object
            properties:
                shops:
                    type: array
                    items:
                        $ref: '#/components/schemas/shop.v1.ShopInfo'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        shop.v1.ShopInfo:
            type: object
            properties:
                id:
                    type: string
                shopName:
                    type: string
                description:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 店铺实体
        shop.v1.UpdateShopRequest:
            type: object
            properties:
                id:
                    type: string
                shopName:
                    type: string
                description:
                    type: string
            description: 更新店铺
        shop.v1.UpdateShopResponse:
            type: object
            properties:
                shop:
                    $ref: '#/components/schemas/shop.v1.ShopInfo'
...